Reorganizados arquivos da interface do usuario para ficar melhor dividido
